% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_titles.R
\name{generate_titles}
\alias{generate_titles}
\title{Generate Fast & Furious inspired movie titles}
\usage{
generate_titles(first_word = "Fast", second_word = "Furious")
}
\arguments{
\item{first_word}{The first root word of the movie titles. Defaults to \code{"Fast"}.}

\item{second_word}{The second root word of the movie titles. Defaults to \code{"Furious"}.}
}
\description{
Generate a family of movie titles based on a series all about family. Takes
two words as arguments to create the titles.
}
\examples{
generate_titles()
generate_titles("Star", "Wars")
}
